requirements.txt
ConfigArgParse==1.5.3
imageio==2.25.0
kornia==0.6.9
lpips==0.1.4
matplotlib==3.6.3
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.4.0
plyfile==0.7.4
scikit_image==0.19.3
scipy==1.11.1
torch_efficient_distloss==0.1.3
tqdm==4.64.1
wandb==0.13.9
